Trigger both variants of nvidia(-tesla)-alternative
If the -support package is installed for the first time at the
same time as the -alternative package, then there is a window of
opportunity where no trigger is generated and thus no modprobe.d
alternatives get installed. Explicitly force the trigger when we
are configured to close the race.